Database-HW02 copy/requirements.txt
Flask==3.0.3
mysql-connector==2.2.9
mysql-connector-python==9.0.0
